requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
selenium>=4.8.0
Pillow>=9.4.0
lxml>=4.9.0